Test SDK Tracer
Move the tracer type and related functionality to its own file to help
discoverability.

Test the tracer type including its concurrent safety (part of #1297).
